Skip trying to resolve the templateLock attribute
The templateLock attribute is an enum with the possible string values "all" and "insert" as well as a third boolean value false.

The type property on this attribute contains an array which specifies both "string" and "boolean" to reflect the possible enum values.

Multiple types on a single attribute is however not currently supported in this library and is therefore skipped here.

# Conflicts:
#	src/Schema/Types/BlockTypes.php